+++ arbitragelab/__init__.py
@@ -6,3 +6,4 @@ reproducible, interpretable, and easy to use tools.
 import arbitragelab.cointegration_approach as cointegration_approach
 import arbitragelab.distance_approach as distance_approach
 import arbitragelab.other_approaches as other_approaches
+import arbitragelab.optimal_mean_reversion as optimal_mean_reversion